+++ tests/test_revoke.py
@@ -1,5 +1,5 @@
 import pytest
-
+import util
 
 def test_revoke_strategy_from_vault(
         token, vault, strategy, amount, user, gov, RELATIVE_APPROX
@@ -29,3 +29,32 @@ def test_revoke_strategy_from_strategy(
     strategy.setEmergencyExit()
     strategy.harvest()
     assert pytest.approx(token.balanceOf(vault.address), rel=RELATIVE_APPROX) == amount
+
+def test_emergency_exit(
+        accounts, token, vault, strategy, user, strategist, amount, RELATIVE_APPROX, chain
+):
+    # Deposit to the vault
+    token.approve(vault.address, amount, {"from": user})
+    vault.deposit(amount, {"from": user})
+    strategy.setBorrowLimit(100 * 1e18)
+    print(f"vault pps: {vault.pricePerShare() / 1e18}")
+    strategy.harvest()
+    assert pytest.approx(strategy.estimatedTotalAssets(), rel=RELATIVE_APPROX) == amount
+    util.stateOfVault(vault, strategy, token)
+
+    chain.sleep(3600 * 6)  # 6 hrs for pps to recover
+    chain.mine(1)
+
+    # set emergency and exit
+    strategy.setEmergencyExit({"from": strategist})
+    print(f"\n before harvest")
+    print(f"vault pps: {vault.pricePerShare()}")
+    util.stateOfStrat(strategy, token)
+    strategy.harvest({"from": strategist})
+    print(f"\n after harvest")
+    print(f"vault pps: {vault.pricePerShare()}")
+    util.stateOfStrat(strategy, token)
+    # dust
+    assert strategy.estimatedTotalAssets() < 1e16
+
+    util.stateOfVault(vault, strategy, token)
